Refactor test & move to test folder
